move get and get_by_name to machine_groups file
