+++ inductiva/resources/machine_groups.py
@@ -822,3 +822,54 @@ class MPICluster(BaseMachineGroup):
 
     def _log_estimated_spot_vm_savings(self) -> None:
         return
+    
+
+def _fetch_machine_groups_from_api():
+    """Get all active machine groups of a user from the API."""
+    try:
+        api = compute_api.ComputeApi(inductiva.api.get_client())
+        response = api.list_active_user_instance_groups()
+
+        return response.body
+
+    except inductiva.client.ApiException as api_exception:
+        raise api_exception
+    
+
+def _get_machine_group_class(machine_type: str, is_elastic: bool):
+    """Returns the class of the machine group"""
+    if is_elastic:
+        mg_class = ElasticMachineGroup
+    elif machine_type == "standard":
+        mg_class = MachineGroup
+    elif machine_type == "mpi":
+        mg_class = MPICluster
+    else:
+        raise ValueError("Unknown resource configuration.")
+    return mg_class
+    
+
+def get_by_name(machine_name: str):
+    """Returns the machine group corresponding to `machine_name`."""
+    try:
+        api = compute_api.ComputeApi(inductiva.api.get_client())
+        response = api.get_vm_group_by_name({"name": machine_name}).body
+        mg_class = _get_machine_group_class(response["type"],
+                                            response["is_elastic"])
+        return mg_class.from_api_response(response)
+    except inductiva.client.ApiException as api_exception:
+        raise api_exception
+
+
+def get():
+    """Returns a list of 'Resource' objects."""
+
+    # Retrive the active resource names
+    machine_groups = _fetch_machine_groups_from_api()
+    machine_group_list = []
+
+    for mg in machine_groups:
+        mg_class = _get_machine_group_class(mg["type"], mg["is_elastic"])
+        machine_group_list.append(mg_class.from_api_response(mg))
+
+    return machine_group_list
