+++ inductiva/resources/utils.py
@@ -174,54 +174,3 @@ def get_machine_dict(machines):
         table["Running task"].append(task_id)
 
     return table
-
-
-def _fetch_machine_groups_from_api():
-    """Get all active machine groups of a user from the API."""
-    try:
-        api = compute_api.ComputeApi(inductiva.api.get_client())
-        response = api.list_active_user_instance_groups()
-
-        return response.body
-
-    except inductiva.client.ApiException as api_exception:
-        raise api_exception
-
-
-def get_by_name(machine_name: str):
-    """Returns the machine group corresponding to `machine_name`."""
-    try:
-        api = compute_api.ComputeApi(inductiva.api.get_client())
-        response = api.get_vm_group_by_name({"name": machine_name}).body
-        mg_class = _get_machine_group_class(response["type"],
-                                            response["is_elastic"])
-        return mg_class.from_api_response(response)
-    except inductiva.client.ApiException as api_exception:
-        raise api_exception
-
-
-def _get_machine_group_class(machine_type: str, is_elastic: bool):
-    """Returns the class of the machine group"""
-    if is_elastic:
-        mg_class = resources.ElasticMachineGroup
-    elif machine_type == "standard":
-        mg_class = resources.MachineGroup
-    elif machine_type == "mpi":
-        mg_class = resources.MPICluster
-    else:
-        raise ValueError("Unknown resource configuration.")
-    return mg_class
-
-
-def get():
-    """Returns a list of 'Resource' objects."""
-
-    # Retrive the active resource names
-    machine_groups = _fetch_machine_groups_from_api()
-    machine_group_list = []
-
-    for mg in machine_groups:
-        mg_class = _get_machine_group_class(mg["type"], mg["is_elastic"])
-        machine_group_list.append(mg_class.from_api_response(mg))
-
-    return machine_group_list
