+++ tutorials/how_to/use-local-task-runner.md
@@ -99,7 +99,7 @@ Code Example:
 import inductiva
 
 # Specify the machine group
-machine_group = inductiva.resources.get_by_name(
+machine_group = inductiva.resources.machine_groups.get_by_name(
     machine_name="<machine-group-name>")
 
 # Download the input files
